feat(ContextMenu): Pulled out contextmenu2 from treeviewer for reuse.
