+++ src/common/components/ContextMenu.less
@@ -1,0 +1,15 @@
+@import "../styles/variables.less";
+
+.context-menu {
+	display: flex;
+	flex-grow: 1;
+	flex-direction: column;
+	border-radius: @border-radius;
+}
+
+.bp4-popover2 .bp4-popover2-content,
+.bp4-menu {
+	background: @container-fg-color-bright;
+	color: @text-color;
+	border-radius: @border-radius;
+}
