+++ src/common/components/ContextMenu.tsx
@@ -1,4 +1,6 @@
-import React from "react";
+import "./ContextMenu.less";
+
+import React, { FC, PropsWithChildren } from "react";
 import {
 	IconName,
 	MenuDivider,
@@ -6,14 +8,16 @@ import {
 	Menu,
 	Intent,
 } from "@blueprintjs/core";
+import { ContextMenu2 } from "@blueprintjs/popover2";
 
 export interface CreateMenuItem {
-	item?: string;
+	item?: string | React.ReactNode;
 	key?: string;
 	cb?: (key?: string) => void;
 	icon?: IconName;
 	intent?: Intent;
 	divider?: boolean;
+	disabled?: boolean;
 	submenu?: CreateMenuItem[];
 }
 export const createContextMenuItems = (items: CreateMenuItem[]) => (
@@ -23,8 +27,7 @@ export const createContextMenuItems = (items: CreateMenuItem[]) => (
 				<MenuDivider key={menuItem.key + "_idx_" + idx} />
 			) : menuItem.submenu ? (
 				<MenuItem
-					intent={menuItem.intent}
-					icon={menuItem.icon}
+					{...menuItem}
 					key={menuItem.key + "_idx_" + idx}
 					text={menuItem.item}
 					onClick={() => menuItem.cb && menuItem.cb(menuItem.key)}
@@ -33,8 +36,7 @@ export const createContextMenuItems = (items: CreateMenuItem[]) => (
 				</MenuItem>
 			) : (
 				<MenuItem
-					intent={menuItem.intent}
-					icon={menuItem.icon}
+					{...menuItem}
 					key={menuItem.key + "_idx_" + idx}
 					text={menuItem.item}
 					onClick={() => menuItem.cb && menuItem.cb(menuItem.key)}
@@ -43,3 +45,19 @@ export const createContextMenuItems = (items: CreateMenuItem[]) => (
 		)}
 	</Menu>
 );
+
+export const ContextMenu: FC<PropsWithChildren<{ items: CreateMenuItem[] }>> = (
+	props
+) => {
+	const { items, children } = props;
+	return (
+		<ContextMenu2
+			popoverProps={{
+				popoverClassName: "context-menu",
+			}}
+			content={createContextMenuItems(items)}
+		>
+			{children}
+		</ContextMenu2>
+	);
+};
