+++ src/components/panes/editor/result-viewer/styles.less
@@ -105,11 +105,6 @@
 			}
 		}
 
-		.context-menu {
-			display: flex;
-			flex-grow: 1;
-		}
-
 		.content-row {
 			@content-row-modified-width: 30px;
